Moved epileptor_models from tvb_api to base. Cleaned up tvb_api package.
