+++ tvb_epilepsy/base/epileptor_models.py
