+++ tvb_epilepsy/service/epileptor_model_factory.py
@@ -11,7 +11,7 @@ import numpy
 from tvb.simulator.models import Epileptor
 
 from tvb_epilepsy.service.simulator.simulator_custom import EpileptorModel
-from tvb_epilepsy.tvb_api.epileptor_models import EpileptorDP2D, EpileptorDP, EpileptorDPrealistic
+from tvb_epilepsy.base.epileptor_models import EpileptorDP2D, EpileptorDP, EpileptorDPrealistic
 
 AVAILABLE_DYNAMICAL_MODELS = (Epileptor, EpileptorModel, EpileptorDP2D, EpileptorDP, EpileptorDPrealistic)
